fix: core correctness, pooling, and global-init robustness
- shim.c: read Int64 then cast to long so Windows (32-bit long) no longer
  truncates numeric options; matches C# Marshal.WriteInt64 width
- CurlEasyWrapper: stack-pointer for SetLongOption (drops a heap alloc per
  numeric option), CoTaskMemUTF8 for strings, zero _handle after cleanup,
  guard Handle getter against use-after-dispose
- HeaderCallback: clear accumulated headers on each HTTP/ status line so
  redirect / proxy-CONNECT / 100-continue hops don't bleed into the result
- CurlEventLoop: build response before MultiRemoveHandle, then TrySetResult;
  block on the work queue when idle instead of 50ms spin
- CurlHandlePool: track handles in a ConcurrentDictionary and remove on
  Discard (fixes unbounded tracker growth)
- CurlGlobal: refcounted init/cleanup under a Lock (won't cleanup-while-in-use)
- RequestMapper: PUT/DELETE/PATCH/OPTIONS via CUSTOMREQUEST; explicit HTTPGET
  on GET to clear residual POST state on pooled handles
- CurlHandler: optional UseSharedEventLoop to share one loop thread

+++ native/shim.c
@@ -2,7 +2,7 @@
 
 int shim_easy_setopt(void* curl, int option, void* param) {
     if (option < CURLOPTTYPE_OBJECTPOINT) {
-        return (int)curl_easy_setopt(curl, (CURLoption)option, *(long*)param);
+        return (int)curl_easy_setopt(curl, (CURLoption)option, (long)(*(int64_t*)param));
     }
     if (option >= CURLOPTTYPE_OFF_T && option < CURLOPTTYPE_BLOB) {
         return (int)curl_easy_setopt(curl, (CURLoption)option, *(curl_off_t*)param);
